requirements.txt
fastapi
uvicorn
requests
streamlit
pandas
pydantic
python-dotenv